requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
SQLAlchemy==2.0.36
Werkzeug==3.1.3
python-dotenv==1.0.1
bcrypt==4.2.1
gunicorn==23.0.0
python-telegram-bot==21.9
google-api-python-client==2.159.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
reportlab==4.2.5
lxml==5.3.0
APScheduler==3.11.0
pdfplumber==0.11.9
